Added JWT, Authentication and Authorization

+++ WebApi/Application/CustomerOperations/Commands/DeleteCustomer/DeleteCustomerCommand.cs
@@ -9,12 +9,14 @@ namespace WebApi.Application.CustomerOperations.Commands.DeleteCustomer
 {
     public class DeleteCustomerCommand
     {
-        private readonly IMovieStoreDbContext _dbContext;
         public int CustomerId { get; set; }
+        private readonly IMovieStoreDbContext _dbContext;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
         public DeleteCustomerCommand(IMovieStoreDbContext dbContext, IHttpContextAccessor httpContextAccessor)
         {
             _dbContext = dbContext;
+            _httpContextAccessor = httpContextAccessor;
         }
 
         public void Handle()
@@ -23,8 +25,15 @@ namespace WebApi.Application.CustomerOperations.Commands.DeleteCustomer
             if (customer is null)
                 throw new InvalidOperationException("Silnecek Müşteri bulunamadı.");
 
+            int requestOwnerId = int.Parse(_httpContextAccessor.HttpContext.User.Claims.SingleOrDefault(claim => claim.Type == "customerId").Value);
+            if (requestOwnerId != customer.Id)
+            {
+                throw new InvalidOperationException("Yalnızca kendi hesabınızı silebilirsiniz.");
+            }
+
             _dbContext.Customers.Remove(customer);
             _dbContext.SaveChanges();
         }
+
     }
 }
